Fixing a rounding issue

+++ A1/part2.py
@@ -9,7 +9,7 @@ from matplotlib import interactive
 def KNearestNeighbours (distances, k):
     topKVals, topKIndices = tf.nn.top_k(tf.negative(distances), k=k)
     N = distances.get_shape().as_list()[1]
-    oneHot = tf.one_hot(topKIndices, depth=N, on_value=tf.to_float(1.0/k))
+    oneHot = tf.one_hot(topKIndices, depth=N, on_value=tf.to_double(1.0/k))
     responsibility = tf.reduce_sum(oneHot, axis=1)
     return responsibility
 
@@ -20,7 +20,7 @@ def calculateMSE(X_train, Y_train, X_test, Y_test, K):
     responsibility = KNearestNeighbours(distances, K)
     Y_head = tf.matmul(responsibility, Y_train)
     MSE = tf.reduce_sum(tf.square(tf.subtract(Y_test, Y_head)))
-    MSE = tf.divide(MSE, tf.to_float(2*N1))
+    MSE = tf.divide(MSE, tf.to_double(2*N1))
     return MSE
 
 if __name__ == "__main__":
@@ -33,13 +33,13 @@ if __name__ == "__main__":
     validData, validTarget = Data[randIdx[80:90]], Target[randIdx[80:90]]
     testData, testTarget = Data[randIdx[90:100]], Target[randIdx[90:100]]
 
-    X_train = tf.placeholder(tf.float32, [80, 1])
-    X_valid = tf.placeholder(tf.float32, [10, 1])
-    X_test = tf.placeholder(tf.float32, [10, 1])
+    X_train = tf.placeholder(tf.float64, [80, 1])
+    X_valid = tf.placeholder(tf.float64, [10, 1])
+    X_test = tf.placeholder(tf.float64, [10, 1])
 
-    Y_train = tf.placeholder(tf.float32, [80, 1])
-    Y_valid = tf.placeholder(tf.float32, [10, 1])
-    Y_test = tf.placeholder(tf.float32, [10, 1])
+    Y_train = tf.placeholder(tf.float64, [80, 1])
+    Y_valid = tf.placeholder(tf.float64, [10, 1])
+    Y_test = tf.placeholder(tf.float64, [10, 1])
 
     Ks = [1, 3, 5, 50]
     with tf.Session() as sess:
@@ -57,7 +57,7 @@ if __name__ == "__main__":
     NewData = np.linspace(0.0, 11.0, num=1000) [:, np.newaxis]
     NewTarget = np.sin(NewData) + 0.1 * np.power(NewData, 2) + 0.5 * np.random.randn(1000 , 1)
 
-    X_new_test = tf.placeholder(tf.float32, [1000,1])
+    X_new_test = tf.placeholder(tf.float64, [1000,1])
     Deuc = part1.euclideanDistance(X_new_test, X_train)
     Predictions = dict.fromkeys(Ks)
     with tf.Session() as sess:
@@ -68,6 +68,7 @@ if __name__ == "__main__":
             init = tf.global_variables_initializer()
             sess.run(init)
             Predictions[K] = sess.run(Y_head, {X_train: trainData, X_new_test: NewData, Y_train: trainTarget})
+
     plt.close('all')
     figs = dict.fromkeys(range(1,len(Ks)+1))
     for i, K in enumerate(Ks):
